### 2018-04-25 #702 @whedon
[PRE REVIEW]: fib-tf: A TensorFlow-based Cardiac Electrophysiology Simulator
**Submitting author:** @siravan (<a href="http://orcid.org/0000-0003-2132-1543">Shahriar Iravanian</a>)
**Repository:** <a href="https://github.com/siravan/fib_tf" target ="_blank">https://github.com/siravan/fib_tf</a>
**Version:** v1.0.0
