### 2018-04-30 #702 @arfon
👋 @mgalloy @leios - would you be willing to review this submission for JOSS?
