### 2018-05-07 #719 @mgalloy
I believe the DOI for the "Reconstruction of the action potential of ventricular myocardial fibres" article in your References is "https://doi.org/10.1113/jphysiol.1977.sp011853". I don't see a DOI for the TensorFlow reference.
