### 2018-05-08 #719 @siravan
Thanks for the comments. The shebangs are fixed. I also added the doi you mentioned. The TensorFlow whitepaper does not seem to have a doi, as it is posted on the TensorFlow website.
