### 2018-05-08 #719 @mgalloy
@siravan A couple questions:

- Are you going to make a release via GitHub? Currently, there are no releases.
- Do you have any guidelines for contributing? asking for support? If not, a short section in your README could probably satisfy the "Community guidelines" above.
