### 2018-05-10 #719 @siravan
Yes, I plan to make a release by the end of the review process, when things are more stable.

Thanks for pointing to the lack of contributing guidelines in the README. I will add one soon.

I welcome contributions from a diverse background, as the man goal of fib-tf is experimentation with different ideas regarding solving ODEs with TensorFlow.
