### 2018-05-10 161a241
Update CONTRIBUTING.md
