+++ CONTRIBUTING.md
@@ -1,1 +1,17 @@
+# Contributing
 
+**fib_tf** is a Python framework developed on the top of [TensorFlow](http://tensorflow.org) for 2D cardiac electrophysiology simulation. Its main goal is to test differnt ideas on how to solve ordinary differntial equations using TensorFlow. As such, the scope of **fib_tf** is not limited to cardiac electrophysiology and modeling ionic models. *We love contributation to **fib_tf** and encourage expanding its usage to other areas and applications*. 
+
+
+
+Fork, then clone the repo from the *master* branch. Install the pre-requisite softwares as described in [README](README.md). 
+
+You can test the installation as:
+
+```python
+python3 fenton.py
+```
+
+To merge back, push to your fork and submit a pull request. A detailed comment on pull requests will be very helpful.
+
+**fib_tf** is under the **MIT** license and any contributing code after the merge will also be under the the **MIT** license.
