### 2018-05-21 6338e22
Added comma at end of line 67 to address paper error
This aims to address the error found during paper generation for JOSS over at https://github.com/openjournals/joss-reviews/issues/719
