+++ paper.bib
@@ -64,7 +64,7 @@ year = {2007}
 @article{Beeler1977,
 abstract = {1. A mathematical model of membrane action potentials of mammalian ventricular myocardial fibres is described. The reconstruction model is based as closely as possible on ionic currents which have been measured by the voltage-clamp method.2. Four individual components of ionic current were formulated mathematically in terms of Hodgkin-Huxley type equations. The model incorporates two voltage- and time-dependent inward currents, the excitatory inward sodium current, i(Na), and a secondary or slow inward current, i(s), primarily carried by calcium ions. A time-independent outward potassium current, i(K1), exhibiting inward-going rectification, and a voltage- and time-dependent outward current, i(x1), primarily carried by potassium ions, are further elements of the model.3. The i(Na) is primarily responsible for the rapid upstroke of the action potential, while the other current components determine the configuration of the plateau of the action potential and the re-polarization phase. The relative importance of inactivation of i(s) and of activation of i(x1) for termination of the plateau is evaluated by the model.4. Experimental phenomena like slow recovery of the sodium system from inactivation, frequency dependence of the action potential duration, all-or-nothing re-polarization, membrane oscillations are adequately described by the model.5. Possible inadequacies and shortcomings of the model are discussed.},
 author = {Beeler, G W and Reuter, H},
-doi = {10.1113/jphysiol.1977.sp011853}
+doi = {10.1113/jphysiol.1977.sp011853},
 issn = {0022-3751},
 journal = {The Journal of physiology},
 mendeley-groups = {af{\_}modeling},
