### 2018-05-21 #719 @Kevin-Mattheus-Moerman
@siravan perhaps when you pasted in the DOI you deleted a comma? Not sure if this is it but merge my pull request and we can try `@whedon generate pdf` again, btw you can ask whedon to do this too.
