### 2018-05-21 #719 @Kevin-Mattheus-Moerman
@siravan, in terms of community guidelines and contributing. I would recommend adding a `CONTRIBUTING.md` file and also a `CODE_OF_CONDUCT.md` file.
You can base the latter on the Contributor Covenant, e.g. version 1.4, available at http://contributor-covenant.org/version/1/4

For more information about community guidelines: https://help.github.com/articles/github-community-guidelines/

Also this link offers example CONTRIBUTING.md files from other projects:
https://help.github.com/articles/setting-guidelines-for-repository-contributors/

Let me know if you have questions.
