### 2018-05-21 #719 @Kevin-Mattheus-Moerman
@siravan I see you already added a CONTRIBUTING.md file. I recommend adding a code of conduct file as well.
