+++ court.py
@@ -1,0 +1,504 @@
+#!/usr/bin/env python
+"""
+    A TensorFlow-based 2D Cardiac Electrophysiology Modeler
+
+    Copyright 2017-2018 Shahriar Iravanian (siravan@emory.edu)
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to
+    deal in the Software without restriction, including without limitation the
+    rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+    sell copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in
+    all copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+    FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+    IN THE SOFTWARE.
+"""
+
+import tensorflow as tf
+import numpy as np
+from screen import Screen
+from ionic import IonicModel
+
+class Courtemanche(IonicModel):
+    """
+        The modified Courtemanche atrial model
+    """
+
+    def __init__(self, props):
+        super().__init__(props)
+        self.min_v = -100.0    # mV
+        self.max_v = 50.0     # mV
+
+    def calc_inter(self, V, mod=np):
+        R = 8.3143          # R in membrane (joule/mole_kelvin).
+        T = 310             # T in (kelvin).
+        F = 96.4867         # F in membrane (coulomb/millimole).
+        Cm = 100            #  Cm in membrane (picoF).
+        Na_o = 140          #  Na_o (millimolar).
+        g_K1 = 0.09         #  g_K1 (nanoS/picoF).
+        K_Q10 = 3           #  transient_outward_K_current (dimensionless).
+        g_Kr = 0.029411765  #  rapid_delayed_rectifier_K_current (nanoS/picoF).
+        Ca_o = 1.8          #  (millimolar).
+        I_NaCa_max = 1600   #  Na_Ca_exchanger_current(picoA/picoF).
+        K_mNa = 87.5        #  Na_Ca_exchanger_current (millimolar).
+        K_mCa = 1.38        #  Na_Ca_exchanger_current (millimolar).
+        K_sat = 0.1         #  Na_Ca_exchanger_current (dimensionless).
+        gamma_ = 0.35       #  Na_Ca_exchanger_current (dimensionless).
+        sigma = 1.0         #  sodium_potassium_pump (dimensionless).
+
+        def where(cond, x, y):
+            ret = mod.where(cond, x, y)
+            if type(ret) is np.ndarray and ret.shape == ():
[…]
