+++ courtemanche.h
@@ -1,0 +1,479 @@
+/***********************************************************/
+//  courtemanche.h is part of the Fibulator 3D atrial fibrillation
+//  simulation software
+//  copyright Shahriar Iravanian (siravan@emory.edu) 2016
+//
+//  courtemanche.cu/h implements Courtemanche, Ramirez, Nattel, 1998
+//
+//**********************************************************/
+
+#pragma once
+
+#include "common.h"
+#include "ionic.h"
+
+/*
+   There are a total of 75 entries in the algebraic variable array.
+   There are a total of 21 entries in each of the rate and state variable
+   arrays.
+   There are a total of 49 entries in the constant variable array.
+*/
+
+template <int N>
+__device__ __host__ float Pow(float x) {
+  return powf(x, N);
+}
+
+template <>
+__device__ __host__ inline float Pow<0>(float x) {
+  return 1.0;
+}
+
+template <>
+__device__ __host__ inline float Pow<1>(float x) {
+  return x;
+}
+
+template <>
+__device__ __host__ inline float Pow<2>(float x) {
+  return x * x;
+}
+
+template <>
+__device__ __host__ inline float Pow<3>(float x) {
+  return x * x * x;
+}
+
+template <>
+__device__ __host__ inline float Pow<-1>(float x) {
+  return 1.0 / x;
+}
+
+template <>
+__device__ __host__ inline float Pow<-2>(float x) {
+  return 1.0 / (x * x);
+}
+
+template <>
+__device__ __host__ inline void init_cell<Courtemanche>(float *v, float *w,
+                                                        int stim) {
+  *v = stim ? 20.0 : -81.18;  // state[0] is V in membrane (millivolt).
+  w[0] = 1.117e+01;  // state[_Na_i_] is intracellular_ion_concentrations
+                     // (millimolar).
+  w[1] = 2.98e-3;    // state[_m_] is m in fast_sodium_current_m_gate
+                     // (dimensionless).
+  w[2] = 9.649e-1;   // state[_h_] is h in fast_sodium_current_h_gate
+                     // (dimensionless).
+  w[3] = 9.775e-1;   // state[_j_] is j in fast_sodium_current_j_gate
+                     // (dimensionless).
+  w[4] = 1.39e+02;   // state[_K_i_] is intracellular_ion_concentrations
+                     // (millimolar).
+  w[5] = 3.043e-2;   // state[_oa_] is oa in
+                     // transient_outward_K_current_oa_gate (dimensionless).
+  w[6] = 9.992e-1;   // state[_oi_] is oi in
+                     // transient_outward_K_current_oi_gate (dimensionless).
+  w[7] = 4.966e-3;   // state[_ua_] is ua in
+                     // ultrarapid_delayed_rectifier_K_current_ua_gate
+                     // (dimensionless).
+  w[8] = 9.986e-1;   // state[_ui_] is ui in
+                     // ultrarapid_delayed_rectifier_K_current_ui_gate
[…]
