### 2018-05-21 #719 @siravan
A CODE_OF_CONDUCT.md file is added to the master branch.
