### 2018-05-21 #719 @mgalloy
@Kevin-Mattheus-Moerman  OK, assuming@siravan makes a GitHub release labeled 1.0, then the only thing I haven't found yet is "Functionality documentation", i.e., API documentation. Am I missing it somewhere @siravan?
