### 2018-05-21 cb25bca
multi-rate implemented
