+++ court.py
@@ -37,6 +37,227 @@ class Courtemanche(IonicModel):
         super().__init__(props)
         self.min_v = -100.0    # mV
         self.max_v = 50.0     # mV
+        self.chronic = True
+        self.fast_states = ['V', '_Na_i_', '_m_', '_h_']
+
+    def init_state_variable(self, state, name, value):
+        if name in state:
+            print('Warning! The state variable arlready exists')
+        state[name] = np.full([self.height, self.width], value, dtype=np.float32)
+
+    def define(self, s1=True):
+        """
+            Defines the tensorflow model
+            It sets ode_op, s2_op and V used by other methods
+        """
+        super().define()
+        state = {}
+        self.init_state_variable(state, 'V', -81.18)
+        self.init_state_variable(state, '_Na_i_', 1.117e+01)
+        self.init_state_variable(state, '_m_', 2.98e-3)
+        self.init_state_variable(state, '_h_', 9.649e-1)
+        self.init_state_variable(state, '_j_', 9.775e-1)
+        self.init_state_variable(state, '_K_i_', 1.39e+02)
+        self.init_state_variable(state, '_oa_', 3.043e-2)
+        self.init_state_variable(state, '_oi_', 9.992e-1)
+        self.init_state_variable(state, '_ua_', 4.966e-3)
+        self.init_state_variable(state, '_ui_', 9.986e-1)
+        self.init_state_variable(state, '_xr_', 3.296e-5)
+        self.init_state_variable(state, '_xs_', 1.869e-2)
+        self.init_state_variable(state, '_Ca_i_', 1.013e-4)
+        self.init_state_variable(state, '_d_', 1.367e-4)
+        self.init_state_variable(state, '_f_', 9.996e-1)
+        self.init_state_variable(state, '_f_Ca_', 7.755e-1)
+        self.init_state_variable(state, '_Ca_rel_', 1.488)
+        self.init_state_variable(state, '_u_', 0.0)
+        self.init_state_variable(state, '_v_', 1.0)
+        self.init_state_variable(state, '_w_', 0.9992)
+        self.init_state_variable(state, '_Ca_up_', 1.488)
+
+        # S1 stimulation: vertical along the left side
+        if s1:
+            state['V'][:,:5] = 20.0
+
+        # define the graph...
+        with tf.device('/device:GPU:0'):
+            # Create variables for simulation state
+            State = {}
+            for s in state:
+                State[s] = tf.Variable(state[s])
+
+            State1 = self.solve(State)
+            self.dt_per_step = 1
+
+            fasts = []
+            slows = []
+            for s in State:
+                if s in self.fast_states:
+                    fasts.append(tf.assign(State[s], State1[s]))
+                else:
+                    slows.append(tf.assign(State[s], State1[s]))
+
+            self._ode_op = tf.group(*fasts)
+            self._ops['slow'] = tf.group(*slows)
+            self._V = State['V']  # V is needed by self.image()
+
+
+    def euler(self, g, Rate, dt):
+        return g + Rate * dt
+
+    def δt(self, name):
+        if name in self.fast_states:
+            return self.dt
+        else:
[…]
