### 2018-05-21 d96db71
multi-rate implemented
