### 2018-05-21 505203a
get/set state
