+++ court.py
@@ -45,38 +45,39 @@ class Courtemanche(IonicModel):
             print('Warning! The state variable arlready exists')
         state[name] = np.full([self.height, self.width], value, dtype=np.float32)
 
-    def define(self, s1=True):
+    def define(self, s1=True, state=None):
         """
             Defines the tensorflow model
             It sets ode_op, s2_op and V used by other methods
         """
         super().define()
-        state = {}
-        self.init_state_variable(state, 'V', -81.18)
-        self.init_state_variable(state, '_Na_i_', 1.117e+01)
-        self.init_state_variable(state, '_m_', 2.98e-3)
-        self.init_state_variable(state, '_h_', 9.649e-1)
-        self.init_state_variable(state, '_j_', 9.775e-1)
-        self.init_state_variable(state, '_K_i_', 1.39e+02)
-        self.init_state_variable(state, '_oa_', 3.043e-2)
-        self.init_state_variable(state, '_oi_', 9.992e-1)
-        self.init_state_variable(state, '_ua_', 4.966e-3)
-        self.init_state_variable(state, '_ui_', 9.986e-1)
-        self.init_state_variable(state, '_xr_', 3.296e-5)
-        self.init_state_variable(state, '_xs_', 1.869e-2)
-        self.init_state_variable(state, '_Ca_i_', 1.013e-4)
-        self.init_state_variable(state, '_d_', 1.367e-4)
-        self.init_state_variable(state, '_f_', 9.996e-1)
-        self.init_state_variable(state, '_f_Ca_', 7.755e-1)
-        self.init_state_variable(state, '_Ca_rel_', 1.488)
-        self.init_state_variable(state, '_u_', 0.0)
-        self.init_state_variable(state, '_v_', 1.0)
-        self.init_state_variable(state, '_w_', 0.9992)
-        self.init_state_variable(state, '_Ca_up_', 1.488)
+        if state is None:
+            state = {}
+            self.init_state_variable(state, 'V', -81.18)
+            self.init_state_variable(state, '_Na_i_', 1.117e+01)
+            self.init_state_variable(state, '_m_', 2.98e-3)
+            self.init_state_variable(state, '_h_', 9.649e-1)
+            self.init_state_variable(state, '_j_', 9.775e-1)
+            self.init_state_variable(state, '_K_i_', 1.39e+02)
+            self.init_state_variable(state, '_oa_', 3.043e-2)
+            self.init_state_variable(state, '_oi_', 9.992e-1)
+            self.init_state_variable(state, '_ua_', 4.966e-3)
+            self.init_state_variable(state, '_ui_', 9.986e-1)
+            self.init_state_variable(state, '_xr_', 3.296e-5)
+            self.init_state_variable(state, '_xs_', 1.869e-2)
+            self.init_state_variable(state, '_Ca_i_', 1.013e-4)
+            self.init_state_variable(state, '_d_', 1.367e-4)
+            self.init_state_variable(state, '_f_', 9.996e-1)
+            self.init_state_variable(state, '_f_Ca_', 7.755e-1)
+            self.init_state_variable(state, '_Ca_rel_', 1.488)
+            self.init_state_variable(state, '_u_', 0.0)
+            self.init_state_variable(state, '_v_', 1.0)
+            self.init_state_variable(state, '_w_', 0.9992)
+            self.init_state_variable(state, '_Ca_up_', 1.488)
 
-        # S1 stimulation: vertical along the left side
-        if s1:
-            state['V'][:,:5] = 20.0
[…]
