### 2018-05-23 #719 @leios
@Kevin-Mattheus-Moerman I apologize for taking so long with this review. I ran into a rather annoying build error when attempting to build tensorflow with JIT and CUDA (https://github.com/tensorflow/tensorflow/issues/10220). It took a while for me to find the solution. I should have been quicker to realize that I needed to use another version of `gcc`.

@mgalloy The documentation can be found in the `docs/` directory under `details.html.` It's pretty good. There is also a small link under the Documentation heading in the README. It's not exactly obvious to find.

Ultimately, the software works as described; however (@siravan), I have a few lingering questions that have not been addressed here:

1. The statement of need wasn't quite clear for me. What are the future applications of this software and where has it already been used in an academic (or non-academic) setting? There was a mention of the systems it can be used for, but no mention of what users would be interested in using this software in-particular or any publications that may result from this software. Is it meant to be used by researchers studying stiff ODE systems or is it a test-case to show that tensorflow can provide decently optimized solutions to such systems?

2. There is a mention of "hand-optimized CUDA code" a lot, but I don't see the code anywhere? I am not sure what you are benchmarking against to get "within a factor of 2-3 of..."

3. So far as I can tell, this software uses the forward Euler method, which is known to be unstable and carry a high degree of error for stiff ODE's, which is the explicit purpose of this work. Because of this, the forward Euler method is not typically used in too many physics applications. Instead, most researchers prefer using Runge-Kutta, Crank-Nicholson, or pseudo-spectral methods. I understand that this application is primarily for electrophysiological systems, but it would seem like a good idea to include other methods to solve the ODE systems to be safe. If it is commonplace to use more error-prone methods for these types of simulations, then maybe it is fine not to address this point; however, I know that there are researchers who will not use anything with an error of less than dt^3.

4. Some of the links in the documentation do not work. Those should probably be fixed.

Please let me know if I missed anything!
