### 2018-05-24 bbc2e57
broken links in details are fixed
