### 2018-05-24 adc5b97
python files added to docs
