+++ docs/br.py
@@ -1,0 +1,370 @@
+#!/usr/bin/env python
+"""
+    A TensorFlow-based 2D Cardiac Electrophysiology Modeler
+
+    Copyright 2017-2018 Shahriar Iravanian (siravan@emory.edu)
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to
+    deal in the Software without restriction, including without limitation the
+    rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+    sell copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in
+    all copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+    FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+    IN THE SOFTWARE.
+"""
+
+import tensorflow as tf
+import numpy as np
+from screen import Screen
+from ionic import IonicModel
+
+class BeelerReuter(IonicModel):
+    """
+        The modified 8-variable Beeler-Reuter model:
+
+        Reconstruction of the action potential of ventricular myocardial fibres,
+        Beeler, G.W. and Reuter, H. 1977, Journal of Physiology, 268, 177-210.
+        PubMed ID: 874889
+    """
+
+    def __init__(self, props):
+        super().__init__(props)
+        self.min_v = -90.0    # mV
+        self.max_v = 30.0     # mV
+
+        # note: the first column for d and f gates is multiplies by 2
+        # to reduce the calcium current activation/deactivation gates by
+        # 50% in order to have spiral waves
+        self.ab_coef = np.array(
+                [[0.0005, 0.083,  50.,    0.0,    0.0,    0.057,  1.0],   # ca_x1
+                 [0.0013, -0.06,  20.,    0.0,    0.0,    -0.04,  1.0],   # cb_x1
+                 [0.0000, 0.0,    47.,    -1.0,   47.,    -0.1,   -1.0],  # ca_m
+                 [40.,    -0.056, 72.,    0.0,    0.0,    0.0,    0.0],   # cb_m
+                 [0.126,  -.25,   77.,    0.0,    0.0,    0.0,    0.0],   # ca_h
+                 [1.7,    0.0,    22.5,   0.0,    0.0,    -0.082, 1.0],   # cb_h
+                 [0.055,  -.25,   78.0,   0.0,    0.0,    -0.2,   1.0],   # ca_j
+                 [0.3,    0.0,    32.,    0.0,    0.0,    -0.1,   1.0],   # cb_j
+                 [2*0.095,  -0.01,  -5.,    0.0,    0.0,    -0.072, 1.0],   # ca_d
+                 [2*0.07,   -0.017, 44.,    0.0,    0.0,    0.05,   1.0],   # cb_d
+                 [2*0.012,  -0.008, 28.,    0.0,    0.0,    0.15,   1.0],   # ca_f
[…]

[…]
