### 2018-05-24 c8b91be
file links in details fixed
