+++ docs/screen.py
@@ -1,175 +1,0 @@
-"""
-    A TensorFlow-based 2D Cardiac Electrophysiology Modeler
-    @2017 Shahriar Iravanian (siravan@emory.edu)
-"""
-
-from dll import _bind
-from ctypes import *
-import numpy as np
-
-# in linux, we can simply use the following code instead of importing from dll
-#
-#SDL = CDLL('libSDL2.so')
-#
-#def _bind(fun, argtypes, restype):
-#    f = getattr(SDL, fun, None)
-#    f.argtypes = argtypes
-#    f.restype = restype
-#    return f
-
-SDL_INIT_VIDEO = 0x00000020
-SDL_INIT_EVENTS = 0x00004000
-SDL_WINDOWPOS_UNDEFINED = 0x1fff0000
-SDL_PIXELFORMAT_ARGB8888 = 0x16362004
-SDL_TEXTUREACCESS_STREAMING = 0x00000001
-SDL_QUIT = 0x00000100
-SDL_WINDOWEVENT = 0x00000200
-SDL_WINDOWEVENT_CLOSE = 0x00000020
-
-
-class SDL_WindowEvent(Structure):
-    _fields_ = [("type", c_uint32),
-                ("timestamp", c_uint32),
-                ("windowID", c_uint32),
-                ("event", c_uint8),
-                ("padding1", c_uint8),
-                ("padding2", c_uint8),
-                ("padding3", c_uint8),
-                ("data1", c_int32),
-                ("data2", c_int32),
-                ("padding", c_uint8 * 56)
-                ]
-
-
-class Screen:
-    """
-        A window plotting 2D numpy ndarrays
-    """
-    def __init__(self, height, width, caption=None):
-        """
-            Creates a blank screen
-        """
-        # first, we will declare SDL2 functions we need...
-
-        # int SDL_Init(c_uint32 flags)
-        self._Init = _bind('SDL_Init', [c_uint32], c_int)
-
-        # c_uint32 SDL_WasInit(c_uint32 flags)
-        self._WasInit = _bind('SDL_WasInit', [c_uint32], c_int)
-
-        # SDL_Window* SDL_CreateWindow(const char* title, int x, int y, int w, int h, c_uint32 flags)
-        self._CreateWindow = _bind('SDL_CreateWindow', [c_char_p, c_int, c_int, c_int, c_int, c_uint32], c_void_p)
-
-        # SDL_Renderer* SDL_CreateRenderer(SDL_Window* window,  int index, c_uint32 flags)
-        self._CreateRenderer = _bind('SDL_CreateRenderer', [c_void_p, c_int, c_uint32], c_void_p)
-
-        # SDL_Texture* SDL_CreateTexture(SDL_Renderer* renderer, c_uint32 format, int access, int w, int h)
-        self._CreateTexture = _bind('SDL_CreateTexture', [c_void_p, c_uint32, c_int, c_int, c_int], c_void_p)
-
-        # void SDL_DestroyTexture(SDL_Texture* texture)
-        self._DestroyTexture = _bind('SDL_DestroyTexture', [c_void_p], None)
-
-        # void SDL_DestroyRenderer(SDL_Renderer* renderer)
-        self._DestroyRenderer = _bind('SDL_DestroyRenderer', [c_void_p], None)
-
-        # void SDL_DestroyWindow(SDL_Window* window)
-        self._DestroyWindow = _bind('SDL_DestroyWindow', [c_void_p], None)
-
-        # int SDL_UpdateTexture(SDL_Texture* texture, const SDL_Rect* rect, const void *pixels, int pitch)
[…]
