+++ docs/fenton_jit.py
@@ -1,241 +1,0 @@
-#!/usr/bin/env python
-"""
-    A TensorFlow-based 2D Cardiac Electrophysiology Modeler
-
-    Copyright 2017-2018 Shahriar Iravanian (siravan@emory.edu)
-
-    Permission is hereby granted, free of charge, to any person obtaining a copy
-    of this software and associated documentation files (the "Software"), to
-    deal in the Software without restriction, including without limitation the
-    rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
-    sell copies of the Software, and to permit persons to whom the Software is
-    furnished to do so, subject to the following conditions:
-
-    The above copyright notice and this permission notice shall be included in
-    all copies or substantial portions of the Software.
-
-    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-    FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-    IN THE SOFTWARE.
-"""
-
-import tensorflow as tf
-import numpy as np
-import time
-from screen import Screen
-from tensorflow.python.client import timeline
-
-def make_kernel(a):
-  """Transform a 2D array into a convolution kernel"""
-  a = np.asarray(a)
-  a = a.reshape(list(a.shape) + [1,1])
-  return tf.constant(a, dtype=1)
-
-def simple_conv(x, k):
-  """A simplified 2D convolution operation"""
-  x = tf.expand_dims(tf.expand_dims(x, 0), -1)
-  y = tf.nn.depthwise_conv2d(x, k, [1, 1, 1, 1], padding='SAME')
-  return y[0, :, :, 0]
-
-def laplace(x):
-  """Compute the 2D laplacian of an array"""
-  laplace_k = make_kernel([[0.5, 1.0, 0.5],
-                           [1.0, -6., 1.0],
-                           [0.5, 1.0, 0.5]])
-  return simple_conv(x, laplace_k)
-
-def enforce_boundary(X):
-    """
-        Enforcing the no-flux (Neumann) boundary condition
-    """
-    paddings = tf.constant([[1,1], [1,1]])
-    return tf.pad(X[1:-1,1:-1], paddings, 'SYMMETRIC', name='boundary')
-
-
-class Fenton4vJIT:
-    """
-        The Cherry-Ehrlich-Nattel-Fenton (4v) canine left-atrial model
-
-        Cherry EM, Ehrlich JR, Nattel S, Fenton FH. Pulmonary vein reentry--
-        properties and size matter: insights from a computational analysis.
-        Heart Rhythm. 2007 Dec;4(12):1553-62.
-    """
-
-    def __init__(self, props):
-        for key, val in config.items():
-            setattr(self, key, val)
-        self.min_v = 0.0
-        self.max_v = 1.0
-
-    def differentiate(self, U, V, W, S):
-        """ the state differentiation for the 4v model """
-        # constants for the Fenton 4v left atrial action potential model
-        tau_vp = 3.33
-        tau_vn1 = 19.2
[…]

[…]
