### 2018-05-24 #719 @siravan
@leios Thanks for the comments. I appreciate your comments and persistence in installing the JIT enabled TensorFlow.

1. The initial reason to develop fib_tf was as a test bed for new ideas on cardiac electrophysiology simulation, like adding a new ionic current to a cardiac model. This is actually how we use it in our academic work (no publication yet, but we are working on a paper on enhancing the Courtemanche atrial model). It is much easier to test ideas in a scripting language like python with all its scientific computing ecosystem than in C++. However, fit_tf can also be useful in assessing the feasibility of writing ODE solvers in TensorFlow.

The main power of TensorFlow is its ability to run on multiple GPUs or even on a distributed system with multiple CPUs, each with one or more GPUs. Writing a C++ code to do this is of course extremely difficult. In this situation, the python code could potentially be faster than the single GPU C++.

2. The "hand-optimized CUDA code" mentioned here is the production code that we use for 3D cardiac modeling (for [example](https://aip.scitation.org/doi/10.1063/1.5000350)). My plan is to refactor the C++ code and CUDA kernles and add the relevant parts to fib_tf. The C++ code base in its current form is too complex and has lots of functionality unrelated to fib_tf.

3. Electrophysiology applications are rather unusual in this regard and are commonly solved using the explicit Euler model (or by implicit solvers on occasion). The reason is the extreme stiffness of models. Higher order explicit solvers, such as Runge-Kutta or Linear Multi-Step methods, are not of much help and one still needs a very small time-step for stability. However, the Rush_larsen technique that we have used here is very helpful and allows us to integrate more than half the state variables (the gating variables) by a closed-form integration formula.

Also please note that TensorFlow has an ODE solver (tf.contrib.integrate.odeint), which implements 5th order Runge-Kutta with adaptive step size control and dense output, using the Dormand-Prince method. If it was logical to use a higher order method, we could have used this functionality.

It should be also noted that TensorFlow has an automatic differentiation capability (needed for the machine learning applications). It is possible to implement high order Taylor series methods by using the automatic differentiation (TensorFlow only has backward differentiation, but there is a trick when one can combine two backward differentiations to obtain a forward one needed in the Taylor series). Again this is not a useful technique for cardiac electrophysiology applications, but is potentially applicable to non-stiff models.

4. The links are fixed.
