### 2018-05-24 #719 @leios
@siravan Thanks for detailed response! I believe that answers all my questions for now, but will take one final look at the code soon.
