### 2018-05-31 #719 @Kevin-Mattheus-Moerman
@mgalloy @leios thanks for your review!! :rocket: can you report on any remaining issues or advise on whether you deem this submission acceptable?
@leios the _Functionality documentation_ box is unticked. Can you summarize what you feel is missing?
Thanks
