### 2018-05-31 #719 @leios
@Kevin-Mattheus-Moerman All my boxes are ticked and I think this submission is acceptable.

@mgalloy is still missing the *functionality documentation* box, though. I think they had trouble finding the documentation, so I tried to show them where it was. I don't know if they have looked at this thread since then.
