### 2018-05-31 #719 @mgalloy
The documentation pointed out by @leios is a great explanation . I have checked off the *functionality documentation* and agree this submission is acceptable.
