### 2018-05-31 #719 @Kevin-Mattheus-Moerman
@siravan looks like we are nearly there. I suggest a minor change to the first sentence in your paper. Instead of `on the top of the` consider saying `on top of the`. If you agree with this make the change and render a new pdf here using `@whedon generate pdf`.
Once you are done with that step please archive a copy of the **reviewed software** using a service like [Zenodo](http://zenodo.org/) ([see here to automate DOI creation on new releases](https://guides.github.com/activities/citable-code/)), or [figshare](https://figshare.com/), and list the DOI of the archived version here.
